requirements.txt
streamlit 
streamlit-drawable-canvas 
pillow 
pandas 
groq
gtts
opencv-python 
pytesseract 
pillow 
numpy